\name{Ocal-package}
\alias{Ocal-package}
\alias{Ocal}
\docType{package}
\title{
Orthogonal calibration via posterior projections
}
\description{
The package contains functions to implement orthogonal calibration as developed in "Orthogonal calibration via posterior projections with applications to the Schwarzschild model" by Chakraborty et.al. (2024). The article can be found at  https://arxiv.org/abs/2404.03152 .
}
\details{

Contains function for calibration of univariate and multivariate output models.

}
\author{
Antik Chakraborty

Maintainer: Antik Chakraborty
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
